refactor : session_id가 null 체크 추가
